tidy up routes
