+++ src/routes/ssr/[template]/+page.server.ts
@@ -7,7 +7,15 @@ import { getCommit } from '$lib/git';
 import { build } from '$lib/rollup';
 import { getProps } from '$lib/svelte';
 import { writeTemplate } from '$lib/write-template';
-import type { RequestHandler } from './$types';
+import type { PageServerLoad } from './$types';
+
+interface Data {
+	template: string;
+	html: string;
+	css: string;
+	props?: Record<string, string>;
+	description: string;
+}
 
 type Output = {
 	html?: string;
@@ -42,8 +50,8 @@ const prerender = (code: string): Output => {
 const isChunk = (output: OutputChunk | OutputAsset): output is OutputChunk =>
 	output.type === 'chunk';
 
-export const GET: RequestHandler = async ({ params }) => {
-	const template = params.template;
+export const load: PageServerLoad = async ({ params }) => {
+	const { template } = params;
 
 	const dir = `src/templates/ssr/${template}`;
 	const path = `${dir}/index.svelte`;
@@ -94,15 +102,18 @@ export const GET: RequestHandler = async ({ params }) => {
 	const css = [`/* ${stamp} */`, String(ssr.css), styles].join('\n');
 
 	const description = existsSync(`${dir}/README.md`)
-		? marked.parse(readFileSync(`${dir}/README.md`, 'utf-8'))
+		? await marked.parse(readFileSync(`${dir}/README.md`, 'utf-8'))
 		: `<p><em>no description provided</em></p>`;
 
 	writeTemplate(template, 'ssr', html, css);
 
-	return json({
+	const data: Data = {
+		template,
 		html,
 		css,
 		props,
 		description,
-	});
+	};
+
+	return data;
 };
